==> 04_calling-rates.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Calling rates

In this script, we assess if there is an association between calling rates and abundances of species across sites and years.

## Load necessary libraries  
```{r}
library(tidyverse)
library(dplyr)
library(stringr)
library(vegan)
library(scico)
library(data.table)
library(extrafont)
library(sf)
library(raster)
library(scales)
library(ggplot2)
library(ggspatial)
library(colorspace)
library(lubridate)
library(naniar)
library(ggstatsplot)
library(patchwork)
```

## Load dataframe containing point count and acoustic data
```{r}
datSubset <- read.csv("results/datSubset.csv")
```

## Creating a single dataframe prior to running statistical analysis

```{r}
## filter out species that were only detected in one approach and not in the other

## Nine species were detected only in point count data but not in the acoustic data and can be ignored for downstream analysis: Herring Gull, Northern Cardinal, American Goshawk, Swamp Sparrow, Red-bellied Woodpecker, Pine Siskin, Savannah Sparrow, Eastern Kingbird, Brown-headed Cowbird
## we will remove the above list of species from the point count data
point_count <- datSubset %>%
  filter(data_type == "point_count") %>%
  filter(!common_name %in% c("Herring Gull", 
                             "Northern Cardinal", "American Goshawk", 
                             "Swamp Sparrow", "Red-bellied Woodpecker", 
                             "Pine Siskin",
                             "Savannah Sparrow", "Eastern Kingbird", 
                             "Brown-headed Cowbird"
                             ))

## Six species were detected only in acoustic data and not in the point count data and can also be ignored for downstream analysis: Cape May Warbler, Gray Catbird, Mallard, Prairie Warbler, House Wren, Bobolink
acoustic_data <- datSubset %>%
  filter(data_type == "acoustic_data") %>%
  filter(!common_name %in% c("Cape May Warbler", "Gray Catbird", 
                             "Mallard", "Prairie Warbler", 
                             "House Wren", "Bobolink"
                             ))

# estimate abundance for each species for each site/visit combination
# ignore distance for now as we can model it in future analysis
abundance <- point_count %>%
  group_by(site_id, site_name, year, visit_number,
           scientific_name,
           common_name) %>% 
  summarise(abundance = sum(abundance)) %>%
  ungroup()

# estimate detections for each species for each site and year combination using the manual annotations of the acoustic data
detections <- acoustic_data %>%
  group_by(site_id, site_name, year, visit_number,
           scientific_name,
           common_name) %>% 
  summarise(detections = n()) %>%
  ungroup()

# estimating detetions per minute
# create a table with detection counts per minute per species
detections_per_minute <- acoustic_data %>%
  mutate(minute = floor_date(as.POSIXct(begin_clock_time, format = "%H:%M:%S"), "minute")) %>%
  group_by(site_id, site_name, year, visit_number,
    scientific_name, common_name, 
    minute) %>%
  summarise(detections_per_minute = n(), .groups = 'drop')

# create a single dataframe
data <- full_join(abundance, detections_per_minute) %>%
  replace_na(list(abundance = 0, detections_per_minute = 0))

# create a dataframe for detections pooled (not at the minute level)
data_pooled <- full_join(abundance, detections) %>%
  replace_na(list(abundance = 0, detections = 0))
```

## Detections per minute vs. abundance

Here, we explore how detections per minute vary by species across a range of abundances across each species

```{r}
## using a random cutoff of a minimum of ten individuals in point count
## grouping by species, year (2022 or 2023) since vocal activity may vary by season and site_name (regional variation as a function of site/habitat)
spp_subset <-  data_pooled %>%
  group_by(common_name, year, site_name) %>%
  summarise(abundance = sum(abundance), detections = sum(detections)) %>%
  ungroup() %>%
  filter(abundance >=10 & detections >= 10)

## keeping only the subset of species
data <- data %>%
  semi_join(spp_subset, by = c("common_name", "year", "site_name"))

# create groups of abundance of 0, 1, 2, 3, and 4 & greater
data <- data %>%
  mutate(abundance_group = case_when(abundance == 0 ~ "Zero",
                                     abundance == 1 ~ "One",
                                     abundance == 2 ~ "Two",
                                     abundance == 3 ~ "Three",
                                     abundance >=4 ~ ">=Four"))

# factorize year (commenting out year for now)
# data$year <- factor(data$year)

# relevel abundance group
data$abundance_group <- fct_relevel(data$abundance_group, 
                                    "Zero", "One", "Two", 
                                    "Three", ">=Four")

## visualizations
plots <- list()

for(i in 1:length(unique(data$common_name))){
 
# extract species common name
a <- unique(data$common_name)[i]
  
# subset data for plotting
for_plot <- data[data$common_name==a,]

# visualization
plots[[i]] <- ggplot(for_plot, 
       aes(x = abundance_group, 
           y = detections_per_minute, 
           fill = factor(site_name))) +
  geom_boxplot(alpha = 0.7, width = 0.8, size = 0.3) +  
  scale_fill_manual(
    values = c("#2d1b69", "#548150FF", "#DEB738FF","#852419FF"),
    name = "Site Name"
  ) +
  # facet_wrap(~year, scales = "free") +
  theme_bw() +
  labs(
    x = "\nAbundance",
    y = "Detections per minute\n",
    title = a
  ) +
  theme(
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, 
      face = "bold"
    ),
    axis.text = element_text(
      family = "Century Gothic",
      size = 12
    ),
    axis.text.x = element_text(
      angle = 90, 
      vjust = 0.5,
      hjust = 1
    ),
     legend.title = element_text(
      family = "Century Gothic",
      size = 12,
      face = "bold"
    ),
    legend.text = element_text(
      family = "Century Gothic",
      size = 10
    ),
    plot.title = element_text(
      family = "Century Gothic",
      size = 16,
      face = "bold"
    ),
    strip.text = element_text(
      family = "Century Gothic",
      size = 12
    )
  )
}

# plot and save as a single pdf
cairo_pdf(
  filename = "figs/abundance-groups-detections-per-minute-by-species.pdf",
  width = 12, height = 7,
  onefile = TRUE
)
plots
dev.off() 
```


## Calculating calling rates

Are calling rates consistent across sites and years between species? Here, we explore calling rates further to understand its variation across sites and years. To calculate calling rates, we essentially divide the number of acoustic detections per visit to a site by the number of individuals detected. Note, however that the calculations here are being done at the minute level. In other words, for each minute, what is the calling rate for that species?

```{r}
## calculating calling rate
data <- data %>%
  mutate(calling_rate_per_minute = (detections_per_minute/abundance))

## getting rid of zero values and Inf
data <- data %>%
  filter(calling_rate_per_minute != 0) %>%
  filter(calling_rate_per_minute != Inf)

# figure
fig_callingRate <- ggplot(data, 
                          aes(x = common_name, y = calling_rate_per_minute)) +
  geom_boxplot(fill = "#D29C44FF") +  
  facet_wrap(~site_name, scales ="free") +          
  theme_bw(base_family = "Century Gothic") +    
  theme(
    plot.title = element_text(size = 18, face = "bold"),
    axis.title = element_text(size = 16, face = "bold"),
    axis.text.x = element_text(
      size = 12,
      angle = 45,             
      hjust = 1,              
      face = "italic"         
    ),
    axis.text.y = element_text(size = 14),
    strip.text = element_text(size = 12, face = "bold")
  ) +
  labs(
    x = "Common Name",         
    y = "Calling Rate Per Minute",        
    title = "Calling Rate by Species" 
  )

ggsave(fig_callingRate, filename = "figs/callingRates-by-species.png", width = 24, height = 12, device = png(), units = "in", dpi = 300)
dev.off()
```
